+++ org.gitools.matrix/src/main/java/org/gitools/matrix/transform/Log10Function.java
@@ -1,0 +1,26 @@
+package org.gitools.matrix.transform;
+
+
+import org.gitools.api.matrix.ConfigurableTransformFunction;
+import org.gitools.api.matrix.IMatrixPosition;
+import org.gitools.api.matrix.TransformFunction;
+
+public class Log10Function extends ConfigurableTransformFunction {
+
+    public Log10Function() {
+        super("Log10", "Returns base 10 Log of given value");
+    }
+
+    @Override
+    public Double apply(Double value, IMatrixPosition position) {
+        if (value != null) {
+            return Math.log10(value);
+        }
+        return null;
+    }
+
+    @Override
+    protected void createDefaultParameters() {
+
+    }
+}