+++ org.gitools.matrix/src/main/java/org/gitools/matrix/transform/parameters/DimensionParameter.java
@@ -1,0 +1,12 @@
+package org.gitools.matrix.transform.parameters;
+
+
+import org.gitools.api.matrix.MatrixDimensionKey;
+
+public class DimensionParameter extends AbstractFunctionParameter<MatrixDimensionKey> {
+
+    @Override
+    public boolean validate(MatrixDimensionKey parameter) {
+        return super.validate(parameter);
+    }
+}